+++ kororaa_graphql_api/schema/toshi_hazard/__init__.py
@@ -1,0 +1,4 @@
+from .hazard_schema import GriddedLocation, GriddedLocationResult, ToshiHazardCurveResult
+from .toshi_hazard_rev0 import hazard_curves_dynamodb
+from .toshi_hazard_rev1 import hazard_curves_dataframe
+from .toshi_hazard_rev2 import hazard_curves