+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
-## 0.1.0 (2022-05-31)
+## [0.2.0] - 2022-08-04
+
+### Added
+ * dataframe interim model
+ * query_v3 suupport against HazardAggregation table
+ * gridded_location query
+ * arbitrary location added to hazard_curve resolver
+
+## [0.1.0] - 2022-05-31
 
 * First release on PyPI.